Make things simpler
